StudyRAG/requirements.txt
fastapi
inngest
uvicorn
llama-index-core
llama-index-readers-file
python-dotenv
qdrant-client
streamlit
openai
